pyproject.toml
[project]
name = "rzee"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "blockbuster>=1.5.25",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "uvicorn>=0.35.0",
]
